Narrowing/rounding: 128 documentation-only explicit casts across 35 files (F10163)
Adds explicit casts at 128 of 131 in-scope clang-tidy
bugprone-narrowing-conversions/incorrect-roundings sites, each checked
against the blob's own disassembly for the matching idiom (e.g.
voicedp.c's fildll/fstps rounding-at-store shape) before the cast was
added -- documentation of an existing implicit conversion, not a
behavior change. Three sites in voicedp.c deliberately left alone
(exact reasoning not restated here; see the per-file diffs). Also
fixes a mutation-anchor regression the pass introduced in
v32anstone.json (3 entries) to match GenerateAnsTone's new
(short)count cast.

+++ src/callprog/toneiir.c
@@ -185,7 +185,7 @@ toneiir_create(struct toneiir *st, const struct toneiir_cfg *cfg)
 
 		st->n = 0;
 		st->env_band = 0;
-		st->env_prev = prev;
+		st->env_prev = (short)prev;
 		st->env_in = 0;
 	}
 
@@ -232,7 +232,7 @@ toneiir_reset(struct toneiir *st)
 
 	st->env_band = 0;
 	st->n = 0;
-	st->env_prev = prev;
+	st->env_prev = (short)prev;
 	st->env_in = 0;
 	st->total = 0;
 	st->run = 0;
@@ -317,7 +317,7 @@ toneiir_progress(struct toneiir *st, short sample)
 		st->n = 0;
 		st->env_band = 0;
 		st->env_in = 0;
-		st->env_prev = prev;
+		st->env_prev = (short)prev;
 	}
 
 	return result;
